requirements.txt
py-translate==1.0.3
pyTelegramBotAPI==3.6.6
